All tests passing on mac
